--- docs/src/app/components/raw-code/date-picker-code.txt
@@ -1,35 +1,0 @@
-//Portrait Dialog
-<DatePicker hintText="Portrait Dialog" />
-//Landscape Dialog
-<DatePicker
-  hintText="Landscape Dialog"
-  mode="landscape"/>
-//Controlled Input
-<DatePicker
-  hintText="Controlled Date Input"
-  value={this.state.controlledDate}
-  onChange={this._handleChange} />
-//Ranged Date Picker
-<DatePicker
-  hintText="Ranged Date Picker"
-  autoOk={this.state.autoOk}
-  minDate={this.state.minDate}
-  maxDate={this.state.maxDate}
-  showYearSelector={this.state.showYearSelector} />
-
-<DatePicker
-  hintText="fr version"
-  // Intl is supported by most modern browsers, see http://caniuse.com/#search=intl
-  // for browsers that don't support it use this polyfill https://github.com/andyearnshaw/Intl.js
-  DateTimeFormat={Intl.DateTimeFormat}
-  wordings={{ok: 'OK', cancel: 'Annuler'}}
-  locale="fr" />
-//Inline Picker
-<DatePicker
-  hintText="Inline"
-  container="inline" />
-//Inline Picker (AutoOk)
-<DatePicker
-  hintText="Inline (AutoOk)"
-  container="inline"
-  autoOk={true} />
