+++ src/date-picker/date-picker.jsx
@@ -6,6 +6,7 @@ import DatePickerDialog from './date-picker-dialog';
 import TextField from '../text-field';
 import ThemeManager from '../styles/theme-manager';
 import DefaultRawTheme from '../styles/raw-themes/light-raw-theme';
+import deprecated from '../utils/deprecatedPropType';
 
 const DatePicker = React.createClass({
 
@@ -30,31 +31,129 @@ const DatePicker = React.createClass({
   },
 
   propTypes: {
+    /**
+     * Constructor for time formatting.
+     * Follow this specificaction: ECMAScript Internationalization API 1.0 (ECMA-402).
+     */
     DateTimeFormat: React.PropTypes.func,
+
+    /**
+     * If true, automatically accept and close the picker on select a date.
+     */
     autoOk: React.PropTypes.bool,
+
+    /**
+     * Used to control how the DatePicker will be displayed when a user tries to set a date.
+     * `dialog` (default) displays the DatePicker as a dialog with a modal.
+     * `inline` displays the DatePicker below the input field (similar to auto complete).
+     */
     container: React.PropTypes.oneOf(['dialog', 'inline']),
+
+    /**
+     * This is the initial date value of the component.
+     * If either `value` or `valueLink` is provided they will override this
+     * prop with `value` taking precedence.
+     */
     defaultDate: React.PropTypes.object,
+
+    /**
+     * Disables the year selection in the date picker.
+     */
+    disableYearSelection: React.PropTypes.bool,
+
+    /**
+     * This function is called to format the date to display in the input box.
+     * By default, date objects are formatted to MM/DD/YYYY.
+     */
     formatDate: React.PropTypes.func,
-    hideToolbarYearChange: React.PropTypes.bool,
+
+    /**
+     * Locale used for formatting date. If you are not using the default value, you
+     * have to provide a DateTimeFormat that supports it. You can use Intl.DateTimeFormat
+     * if it's supported by your environment.
+     * https://github.com/andyearnshaw/Intl.js is a good polyfill.
+     */
     locale: React.PropTypes.string,
+
+    /**
+     * The ending of a range of valid dates. The range includes the endDate.
+     * The default value is current date + 100 years.
+     */
     maxDate: React.PropTypes.object,
+
+    /**
+     * The beginning of a range of valid dates. The range includes the startDate.
+     * The default value is current date - 100 years.
+     */
     minDate: React.PropTypes.object,
+
+    /**
+     * Tells the component to display the picker in portrait or landscape mode.
+     */
     mode: React.PropTypes.oneOf(['portrait', 'landscape']),
+
+    /**
+     * Callback function that is fired when the date value changes. Since there
+     * is no particular event associated with the change the first argument
+     * will always be null and the second argument will be the new Date instance.
+     */
     onChange: React.PropTypes.func,
+
+    /**
+     * Fired when the datepicker dialog is dismissed.
+     */
     onDismiss: React.PropTypes.func,
+
+    /**
+     * Callback function that is fired when the datepicker field gains focus.
+     */
     onFocus: React.PropTypes.func,
+
+    /**
+     * Fired when the datepicker dialog is shown.
+     */
     onShow: React.PropTypes.func,
+
+    /**
+     * Called when touch tap event occurs on text-field.
+     */
     onTouchTap: React.PropTypes.func,
+
+    /**
+     * Called during render time of a given day. If this method returns
+     * false the day is disabled otherwise it is displayed normally.
+     */
     shouldDisableDate: React.PropTypes.func,
-    showYearSelector: React.PropTypes.bool,
+
+    /**
+     *  Enables the year selection in the date picker.
+     */
+    showYearSelector: deprecated(React.PropTypes.element,
+          'Instead, use disableYearSelection.'),
 
     /**
      * Override the inline-styles of the root element.
      */
     style: React.PropTypes.object,
+
+    /**
+     * Override the inline-styles of DatePicker's TextField element.
+     */
     textFieldStyle: React.PropTypes.object,
+
+    /**
+     * Sets the date for the Date Picker programmatically.
+     */
     value: React.PropTypes.any,
+
+    /**
+     * Creates a ValueLink with the value of date picker.
+     */
     valueLink: React.PropTypes.object,
+
+    /**
+     * Wordings used inside the button of the dialog.
+     */
     wordings: React.PropTypes.object,
   },
 
@@ -66,7 +165,7 @@ const DatePicker = React.createClass({
     return {
       formatDate: DateTime.format,
       autoOk: false,
-      showYearSelector: false,
+      disableYearSelection: false,
       style: {},
     };
   },
@@ -110,7 +209,7 @@ const DatePicker = React.createClass({
       onFocus,
       onShow,
       onTouchTap,
-      showYearSelector,
+      disableYearSelection,
       style,
       textFieldStyle,
       valueLink,
@@ -140,9 +239,8 @@ const DatePicker = React.createClass({
           minDate={minDate}
           maxDate={maxDate}
           autoOk={autoOk}
-          showYearSelector={showYearSelector}
-          shouldDisableDate={this.props.shouldDisableDate}
-          hideToolbarYearChange={this.props.hideToolbarYearChange} />
+          disableYearSelection={disableYearSelection}
+          shouldDisableDate={this.props.shouldDisableDate}/>
       </div>
 
     );
